Fix: remove load event listener and run gridCreator() globally
Running gridCreator function inside and event listener that is waiting for load or even DOMContentLoaded event leads to some problems.

For example, If I wanted to style the child divs, I had to put the related code inside that event listener, while this is not aligned with project's constraints. The load event will only fire when whole page and resources is loaded, meaning any code related to child divs and written after it's event listener, won't get executed since the whole script needs to be downloaded and executed.
The color changer function must have it's own event listener.