Add the feature of grid changing via a button
Develope the program by writing codes that will make possible to change the grid dimension by using the button.
Write an event listener attached to the button. This handler will initiate a prompt window that would ask user to enter squares per side for the grid. The whole script for changing the grid layout and it's color is wrapped inside a function. The event listener will call that function.