Remove `flex-grow: 1` as it's not needed
The rule: `flex-grow: 1` was written in styles.css with perception that it would determine the height of each child div. This was a misunderstanding.

In this code, the height of each child div is determined by: 1- The equality of rows and columns; 2- Container div having a square size (height = width) and; 3- the flex-basis which determines the width of each child div (fraction of overall width).

But, In the end, The amount of vertical growth for each div is conducted by the omitted `align-items: normal[inital value which acts as strech]` rule. As if it's changed to `align-items: start`, you can see that grid's layout won't be as desired.
So, 1- Container being square; 2- Items getting wrapped; 3- The equality of rows and columns inside the square container; 4- The effect of omitted `align-items: normal` will result in the growth of child divs and therefore being square shaped. So `flex-grow: 1` can be deleted from the codebase.