Fix the bug of valid input after an invalid one
If user attempts to change grid layout and enters an invalid value, i.e. input>100, and then enters a valid, the grid doesn't change into the desired layout.

This was because invalid input condition's code body, doesn't end with a return statement so the script keeps on executing with an buggy square per side. Add return statement after gridChanger and it's fixed.