Reduce script size by puttng grid colorings in a function
In line with "DRY!" delete a chunk of code that gets repeated three times. The part of the code that selects child divs and styles them by iterating via forEach is moved into a function.
So whenever it's needed to add rgb random colors and opacity style to the grid, this function is called instead of repeating the code.